Count an E.C.M. burst in ticks, not updates
TimeECM ran down once per update, so a burst that should last thirty-two of
the game's ticks lasted a quarter as long at sixty frames a second - and,
when it was ours to run, cost a quarter of the energy with it.

It was missed by the animations item because no golden scenario fires an
E.C.M.: none of the five arms a missile, so nothing ever ran the code. Hence
two tests that reach it directly, one for the length of the burst and one
for what it costs, both run at a quarter of a tick an update.

The countdown takes the smaller of the tick and what is left, so a part-tick
step cannot carry it past zero - the same clamp the other dwell times got.

+++ src/elite/libs/EliteSharpLib/Ships/PlayerShip.cs
@@ -117,7 +117,11 @@ internal sealed class PlayerShip
 
     internal float Climb { get; set; }
 
-    internal int EcmActive { get; set; }
+    /// <summary>
+    /// Gets or sets how much longer an E.C.M. burst runs, counted in the
+    /// game's own ticks. Zero when none is running.
+    /// </summary>
+    internal float EcmActive { get; set; }
 
     /// <summary>
     /// Gets or sets the energy banks, between <see cref="EnergyMin"/> and
